Moved some more tests around items, credit, subjects and genres
