+++ catalogue_api/api/src/test/scala/uk/ac/wellcome/platform/api/works/ApiWorksTest.scala
@@ -171,96 +171,6 @@ class ApiWorksTest extends ApiWorksTestBase {
     }
   }
 
-  it("includes 'items' if the items include is present, even with no items") {
-    val work = workWith(
-      canonicalId = "dgdb712",
-      title = "Without windows or wind or washing-up liquid",
-      items = List()
-    )
-    insertIntoElasticSearch(work)
-
-    eventually {
-      server.httpGet(
-        path = s"/$apiPrefix/works/${work.canonicalId}?includes=items",
-        andExpect = Status.Ok,
-        withJsonBody = s"""
-                          |{
-                          | "@context": "https://localhost:8888/$apiPrefix/context.json",
-                          | "type": "Work",
-                          | "id": "${work.canonicalId}",
-                          | "title": "${work.title.get}",
-                          | "creators": [ ],
-                          | "items": [ ],
-                          | "subjects": [ ],
-                          | "genres": [ ],
-                          | "publishers": [ ],
-                          | "placesOfPublication": [ ]
-                          |}
-          """.stripMargin
-      )
-    }
-  }
-
-  it("includes credit information in API responses") {
-    val location = DigitalLocation(
-      locationType = "thumbnail-image",
-      url = "",
-      credit = Some("Wellcome Collection"),
-      license = License_CCBY
-    )
-    val item = IdentifiedItem(
-      canonicalId = "chu27a8",
-      sourceIdentifier = sourceIdentifier,
-      identifiers = List(),
-      locations = List(location)
-    )
-    val workWithCopyright = IdentifiedWork(
-      title = Some("A scarf on a squirrel"),
-      sourceIdentifier = sourceIdentifier,
-      version = 1,
-      canonicalId = "yxh928a",
-      items = List(item))
-    insertIntoElasticSearch(workWithCopyright)
-
-    eventually {
-      server.httpGet(
-        path = s"/$apiPrefix/works?includes=items",
-        andExpect = Status.Ok,
-        withJsonBody = s"""
-          |{
-          |  ${resultList()},
-          |  "results": [
-          |   {
-          |     "type": "Work",
-          |     "id": "${workWithCopyright.canonicalId}",
-          |     "title": "${workWithCopyright.title.get}",
-          |     "creators": [ ],
-          |     "subjects": [ ],
-          |     "genres": [ ],
-          |     "publishers": [ ],
-          |     "placesOfPublication": [ ],
-          |     "items": [
-          |       {
-          |         "id": "${item.canonicalId}",
-          |         "type": "${item.ontologyType}",
-          |         "locations": [
-          |           {
-          |             "type": "${location.ontologyType}",
-          |             "url": "",
-          |             "locationType": "${location.locationType}",
-          |             "license": ${license(location.license)},
-          |             "credit": "${location.credit.get}"
-          |           }
-          |         ]
-          |       }
-          |     ]
-          |   }
-          |  ]
-          |}""".stripMargin
-      )
-    }
-  }
-
   it(
     "returns the requested page of results when requested with page & pageSize, alongside correct next/prev links ") {
     val works = createWorks(3)
@@ -524,76 +434,7 @@ class ApiWorksTest extends ApiWorksTestBase {
     }
   }
 
-  it("includes subject information in API responses") {
-    val workWithSubjects = IdentifiedWork(
-      title = Some("A seal selling seaweed sandwiches in Scotland"),
-      sourceIdentifier = sourceIdentifier,
-      version = 1,
-      identifiers = List(),
-      canonicalId = "test_subject1",
-      subjects = List(Concept("fish"), Concept("gardening"))
-    )
-    insertIntoElasticSearch(workWithSubjects)
 
-    eventually {
-      server.httpGet(
-        path = s"/$apiPrefix/works",
-        andExpect = Status.Ok,
-        withJsonBody = s"""
-                          |{
-                          |  ${resultList()},
-                          |  "results": [
-                          |   {
-                          |     "type": "Work",
-                          |     "id": "${workWithSubjects.canonicalId}",
-                          |     "title": "${workWithSubjects.title.get}",
-                          |     "creators": [],
-                          |     "subjects": [ ${concepts(
-                            workWithSubjects.subjects)} ],
-                          |     "genres": [ ],
-                          |     "publishers": [ ],
-                          |     "placesOfPublication": [ ]
-                          |   }
-                          |  ]
-                          |}""".stripMargin
-      )
-    }
-  }
-
-  it("includes genre information in API responses") {
-    val workWithSubjects = IdentifiedWork(
-      title = Some("A guppy in a greenhouse"),
-      sourceIdentifier = sourceIdentifier,
-      version = 1,
-      identifiers = List(),
-      canonicalId = "test_subject1",
-      genres = List(Concept("woodwork"), Concept("etching"))
-    )
-    insertIntoElasticSearch(workWithSubjects)
-
-    eventually {
-      server.httpGet(
-        path = s"/$apiPrefix/works",
-        andExpect = Status.Ok,
-        withJsonBody = s"""
-                          |{
-                          |  ${resultList()},
-                          |  "results": [
-                          |   {
-                          |     "type": "Work",
-                          |     "id": "${workWithSubjects.canonicalId}",
-                          |     "title": "${workWithSubjects.title.get}",
-                          |     "creators": [],
-                          |     "subjects": [ ],
-                          |     "genres": [ ${concepts(workWithSubjects.genres)} ],
-                          |     "publishers": [ ],
-                          |     "placesOfPublication": [ ]
-                          |   }
-                          |  ]
-                          |}""".stripMargin
-      )
-    }
-  }
 
   it(
     "includes a list of identifiers on a list endpoint if we pass ?includes=identifiers") {
